// front/src/app/api-client/model/ingredient-dto.ts
/**
 * Agenda familial - API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Ingrédient réutilisable entre recettes
 */
export interface IngredientDTO { 
    id?: number;
    nom?: string;
    /**
     * Rayon du magasin (optionnel)
     */
    rayon?: IngredientDTO.RayonEnum | null;
}
export namespace IngredientDTO {
    export const RayonEnum = {
        FruitsLegumes: 'FRUITS_LEGUMES',
        BoucheriePoissonnerie: 'BOUCHERIE_POISSONNERIE',
        Cremerie: 'CREMERIE',
        Epicerie: 'EPICERIE',
        Surgeles: 'SURGELES',
        Boulangerie: 'BOULANGERIE',
        Boissons: 'BOISSONS',
        Autre: 'AUTRE',
    } as const;
    export type RayonEnum = typeof RayonEnum[keyof typeof RayonEnum];
}