// front/src/app/api-client/model/recipe-dto.ts
/**
 * Agenda familial - API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RecipeIngredientDTO } from './recipe-ingredient-dto';


/**
 * Recette type
 */
export interface RecipeDTO { 
    id?: number;
    nom?: string;
    description?: string | null;
    /**
     * Portions pour lesquelles les quantités sont données
     */
    portions?: number;
    /**
     * En minutes
     */
    tempsPreparation?: number | null;
    instructions?: string | null;
    /**
     * Dans l\'ordre d\'affichage
     */
    ingredients?: Array<RecipeIngredientDTO>;
}