// front/src/app/api-client/model/recipe-request.ts
/**
 * Agenda familial - API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RecipeIngredientRequest } from './recipe-ingredient-request';


/**
 * Création / modification d\'une recette type
 */
export interface RecipeRequest { 
    nom: string;
    /**
     * Description courte
     */
    description?: string;
    /**
     * Portions pour lesquelles les quantités sont données
     */
    portions: number;
    /**
     * Temps de préparation en minutes (optionnel)
     */
    tempsPreparation?: number;
    /**
     * Instructions, texte libre multi-lignes
     */
    instructions?: string;
    /**
     * Ingrédients, dans l\'ordre d\'affichage ; absent = aucun
     */
    ingredients?: Array<RecipeIngredientRequest>;
}